Make a base disability comp controller move itf verification into a concern
