+++ modules/claims_api/app/controllers/claims_api/base_form_controller.rb
@@ -7,9 +7,6 @@ module ClaimsApi
   class BaseFormController < ClaimsApi::ApplicationController
     before_action :validate_json_schema
 
-    STATSD_VALIDATION_FAIL_KEY = 'api.claims_api.526.validation_fail'
-    STATSD_VALIDATION_FAIL_TYPE_KEY = 'api.claims_api.526.validation_fail_type'
-
     # schema endpoint should be wide open
     skip_before_action :validate_json_schema, only: %i[schema]
     skip_before_action :authenticate, only: %i[schema]
@@ -23,25 +20,6 @@ module ClaimsApi
 
     private
 
-    def verification_itf_expiration
-      active = itf_service.get_active('compensation')
-      if !active['intent_to_file'] || active['intent_to_file'].expiration_date < Time.now.utc
-        error = {
-          errors: [
-            {
-              status: 422,
-              details: 'Intent to File Expiration Date not valid, resubmit ITF.'
-            }
-          ]
-        }
-        render json: error, status: 422
-      end
-    end
-
-    def itf_service
-      EVSS::IntentToFile::Service.new(target_veteran)
-    end
-
     def validate_json_schema
       ClaimsApi::FormSchemas.validate!(self.class::FORM_NUMBER, form_attributes)
     rescue ClaimsApi::JsonApiMissingAttribute => e
@@ -54,32 +32,6 @@ module ClaimsApi
       payload_attributes ||= {}
 
       payload_attributes
-    end
-
-    def valid_526_response
-      {
-        data: {
-          type: 'claims_api_auto_established_claim_validation',
-          attributes: {
-            status: 'valid'
-          }
-        }
-      }.to_json
-    end
-
-    def format_526_errors(errors)
-      errors.map do |error|
-        { status: 422, detail: "#{error['key']} #{error['detail']}", source: error['key'] }
-      end
-    end
-
-    def track_526_validation_errors(errors)
-      StatsD.increment STATSD_VALIDATION_FAIL_KEY
-
-      errors.each do |error|
-        key = error['key'].gsub(/\[(.*?)\]/, '')
-        StatsD.increment STATSD_VALIDATION_FAIL_TYPE_KEY, tags: ["key: #{key}"]
-      end
     end
   end
 end
