+++ modules/claims_api/app/controllers/claims_api/concerns/itf_verification.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module ClaimsApi
+  module ItfVerification
+    extend ActiveSupport::Concern
+
+    included do
+      def foo
+        binding.pry
+      end
+    end
+  end
+end
