+++ modules/claims_api/app/controllers/claims_api/v0/forms/disability_compensation_controller.rb
@@ -1,21 +1,24 @@
 # frozen_string_literal: true
 
-require_dependency 'claims_api/base_form_controller'
+require_dependency 'claims_api/base_disability_compensation_controller'
+require_dependency 'claims_api/concerns/itf_verification'
 require 'jsonapi/parser'
 
 module ClaimsApi
   module V0
     module Forms
-      class DisabilityCompensationController < BaseFormController
+      class DisabilityCompensationController < BaseDisabilityCompensationController
+        include ClaimsApi::ItfVerification
         FORM_NUMBER = '526'
         skip_before_action(:authenticate)
         skip_before_action(:verify_power_of_attorney)
-        before_action :verification_itf_expiration, only: %i[submit_form_526]
+        before_action :verify_itf, only: %i[submit_form_526]
         skip_before_action :validate_json_schema, only: %i[upload_supporting_documents]
         skip_before_action :verify_mvi, only: %i[submit_form_526 validate_form_526]
         skip_before_action :log_request, only: %i[validate_form_526]
 
         def submit_form_526
+          foo
           service = EVSS::DisabilityCompensationForm::ServiceAllClaim.new(auth_headers)
           auto_claim = ClaimsApi::AutoEstablishedClaim.create(
             status: ClaimsApi::AutoEstablishedClaim::PENDING,
